refactor(console): dto를 controller와 view 패키지로 이동
