+++ src/main/java/roomescape/console/FrontController.java
@@ -3,8 +3,8 @@ package roomescape.console;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 import roomescape.console.controller.ConsoleController;
-import roomescape.console.request.ConsoleRequest;
-import roomescape.console.response.ConsoleResponse;
+import roomescape.console.view.ConsoleRequest;
+import roomescape.console.view.ConsoleResponse;
 import roomescape.console.utils.ControllerMapper;
 import roomescape.console.view.InputView;
 import roomescape.console.view.OutputView;
