+++ src/main/java/roomescape/console/controller/ConsoleController.java
@@ -1,7 +1,7 @@
 package roomescape.console.controller;
 
-import roomescape.console.request.ConsoleRequest;
-import roomescape.console.response.ConsoleResponse;
+import roomescape.console.view.ConsoleRequest;
+import roomescape.console.view.ConsoleResponse;
 
 public interface ConsoleController {
     ConsoleResponse dispatch(ConsoleRequest request);
