+++ src/main/java/roomescape/console/view/ConsoleResponse.java
@@ -1,4 +1,6 @@
-package roomescape.console.response;
+package roomescape.console.view;
+
+import roomescape.console.controller.response.ReservationTimeResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
