+++ src/main/java/roomescape/console/view/OutputView.java
@@ -1,7 +1,5 @@
 package roomescape.console.view;
 
-import roomescape.console.response.ConsoleResponse;
-
 public class OutputView {
     public void printResponse(ConsoleResponse dispatch) {
         System.out.println(dispatch.getContent());
